[gist] add tabs for multiple files
