+++ frontend/classic/js/addon/collab/index.ts
@@ -1,3 +1,3 @@
 import '../public-path';
 export { CollabP2P } from './p2p';
-export { Gist } from './gist';
+export { Gist } from './Gist';
